requirements.txt
# MT5_Bot_mk2 dependencies
# Low-spec CPU-first stack

# Core
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# ML — lightweight
scikit-learn>=1.3.0
joblib>=1.3.0

# ZMQ
pyzmq>=25.0.0

# Optional: PyTorch CPU (uncomment ONE of the following)
# torch==2.1.0+cpu  --index-url https://download.pytorch.org/whl/cpu
# torch-directml    (Windows DirectML GPU acceleration)

# Optional: ONNX Runtime (CPU or DirectML)
# onnxruntime>=1.16.0
# onnxruntime-directml>=1.16.0

# LightGBM tabular ensemble component
lightgbm>=4.0.0
